+++ src/main/java/com/manchui/domain/auth/dto/NameDTO.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.dto;
+package com.manchui.domain.auth.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
