+++ src/main/java/com/manchui/domain/auth/service/CustomUserDetailsService.java
@@ -1,8 +1,8 @@
-package com.manchui.domain.service;
+package com.manchui.domain.auth.service;
 
-import com.manchui.domain.dto.CustomUserDetails;
-import com.manchui.domain.entity.User;
-import com.manchui.domain.repository.UserRepository;
+import com.manchui.domain.auth.dto.CustomUserDetails;
+import com.manchui.domain.user.entity.User;
+import com.manchui.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
