+++ src/main/java/com/manchui/domain/auth/service/Oauth2Service.java
@@ -1,11 +1,12 @@
-package com.manchui.domain.service;
+package com.manchui.domain.auth.service;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.manchui.domain.entity.Image;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.service.ImageServiceImpl;
+import com.manchui.domain.user.entity.User;
 import com.manchui.domain.repository.ImageRepository;
-import com.manchui.domain.repository.UserRepository;
+import com.manchui.domain.user.repository.UserRepository;
 import com.manchui.global.exception.CustomException;
 import com.manchui.global.exception.ErrorCode;
 import com.manchui.global.jwt.JWTUtil;
